requirements/base.txt
Django==1.8.6
djangorestframework==3.3.0
psycopg2==2.6.1
Pillow==3.0.0
